chore: move sykmeldt pages to app
