+++ src/app/sykmeldt/[sykmeldtId]/sykmeldinger/page.tsx
@@ -1,20 +1,20 @@
+"use client";
+
 import { PersonIcon } from "@navikt/aksel-icons";
 import { PageContainer, RootPages } from "@navikt/dinesykmeldte-sidemeny";
-import Head from "next/head";
 import type { ReactElement } from "react";
-import { withAuthenticatedPage } from "../../../auth/withAuthentication";
-import PageSideMenu from "../../../components/PageSideMenu/PageSideMenu";
-import PageError from "../../../components/shared/errors/PageError";
-import SykmeldtNotFound from "../../../components/shared/errors/SykmeldtNotFound";
-import PageFallbackLoader from "../../../components/shared/pagefallbackloader/PageFallbackLoader";
-import SykmeldingerList from "../../../components/sykmeldinger/SykmeldingerList";
+import PageSideMenu from "../../../../components/PageSideMenu/PageSideMenu";
+import PageError from "../../../../components/shared/errors/PageError";
+import SykmeldtNotFound from "../../../../components/shared/errors/SykmeldtNotFound";
+import PageFallbackLoader from "../../../../components/shared/pagefallbackloader/PageFallbackLoader";
+import SykmeldingerList from "../../../../components/sykmeldinger/SykmeldingerList";
 import {
   createSykmeldingerBreadcrumbs,
   useUpdateBreadcrumbs,
-} from "../../../hooks/useBreadcrumbs";
-import useFocusRefetch from "../../../hooks/useFocusRefetch";
-import { useSykmeldt } from "../../../hooks/useSykmeldt";
-import { fnrText, formatNameSubjective } from "../../../utils/sykmeldtUtils";
+} from "../../../../hooks/useBreadcrumbs";
+import useFocusRefetch from "../../../../hooks/useFocusRefetch";
+import { useSykmeldt } from "../../../../hooks/useSykmeldt";
+import { fnrText, formatNameSubjective } from "../../../../utils/sykmeldtUtils";
 
 function Sykmeldinger(): ReactElement {
   const { sykmeldtId, sykmeldt, isLoading, error, sykmeldtNotFound, refetch } =
@@ -45,9 +45,9 @@ function Sykmeldinger(): ReactElement {
         )
       }
     >
-      <Head>
+      {/* <Head>
         <title>Sykmeldinger - Dine Sykmeldte - nav.no</title>
-      </Head>
+      </Head> */}
 
       {isLoading && <PageFallbackLoader text="Laster sykmeldinger" />}
       {sykmeldt && (
@@ -63,7 +63,5 @@ function Sykmeldinger(): ReactElement {
     </PageContainer>
   );
 }
-
-export const getServerSideProps = withAuthenticatedPage();
 
 export default Sykmeldinger;
